Update frigate_vision.yaml
Device notification selection updated to allow for ",',` in device names now.
Reworked the LLMVision action as if it timed out, or resources were exhausted (API overused) the automation would hang due to the "single" execution style. Updated to include a timeout action to hopefully continue on timeout to allow for the Automation to end, even with an LLMVision failure.